ObjectFile.jl Documentation

Index

ObjectFile.DynamicLink

ObjectFile.DynamicLinks

ObjectFile.MagicMismatch

ObjectFile.ObjectHandle

ObjectFile.RPath

ObjectFile.RPath

ObjectFile.Section

ObjectFile.SectionRef

ObjectFile.SectionRef

ObjectFile.Sections

ObjectFile.Sections

ObjectFile.Segment

ObjectFile.SegmentRef

ObjectFile.Segments

ObjectFile.Segments

ObjectFile.StrTab

ObjectFile.SymbolRef

ObjectFile.SymbolRef

ObjectFile.Symbols

ObjectFile.Symbols

ObjectFile.SymtabEntry

ObjectFile.fixed_string

Base.eof

Base.findall

Base.findall

Base.findfirst

Base.findfirst

Base.read

Base.read

Base.seek

Base.seekstart

Base.unsafe_string

Base.unsafe_string

ObjectFile.canonical_rpaths

ObjectFile.deref

ObjectFile.deref

ObjectFile.endianness

ObjectFile.find_libraries

ObjectFile.find_library

ObjectFile.find_library

ObjectFile.format_string

ObjectFile.getindex_ref

ObjectFile.handle

ObjectFile.handle

ObjectFile.handle

ObjectFile.handle

ObjectFile.handle

ObjectFile.handle

ObjectFile.header

ObjectFile.iostream

ObjectFile.is64bit

ObjectFile.isdynamic

ObjectFile.isexecutable

ObjectFile.isglobal

ObjectFile.islibrary

ObjectFile.islocal

ObjectFile.isrelocatable

ObjectFile.isundef

ObjectFile.isweak

ObjectFile.mangle_section_name

ObjectFile.mangle_symbol_name

ObjectFile.path

ObjectFile.path

ObjectFile.read_struct

ObjectFile.readmeta

ObjectFile.readmeta

ObjectFile.readmeta

ObjectFile.rpaths

ObjectFile.section_address

ObjectFile.section_header_offset

ObjectFile.section_header_size

ObjectFile.section_header_type

ObjectFile.section_name

ObjectFile.section_number

ObjectFile.section_offset

ObjectFile.section_size

ObjectFile.segment_header_offset

ObjectFile.segment_header_size

ObjectFile.segment_header_type

ObjectFile.segment_name

ObjectFile.segment_number

ObjectFile.show_collection

ObjectFile.startaddr

ObjectFile.strtab_lookup

ObjectFile.symbol_name

ObjectFile.symbol_name

ObjectFile.symbol_number

ObjectFile.symbol_value

ObjectFile.symtab_entry_offset

ObjectFile.symtab_entry_size

ObjectFile.symtab_entry_type

Base.eof — Method

eof(section::SectionRef)

Returns true if the ObjectHandle that this SectionRef refers to has read beyond the current section's extent

Base.findall — Method

findall(sections::Sections, name::String)

Return a list of sections that match the given name.

Base.findall — Method

findall(sections::Sections, name::String)

Return a list of sections that match one of the given names.

Base.findfirst — Method

findfirst(sections::Sections, name::String)

Return the first section that matches the given name.

Base.findfirst — Method

findfirst(sections::Sections, names::Vector{String})

Return the first section that matches on of the given names.

Base.read — Method

read(section::SectionRef)

Read the contents of the section referred to by section, returning a Vector{UInt8}.

Base.read — Method

read(oh::ObjectHandle, section::Section)

Read the contents of the section referred to by section from the given ObjectHandle, returning a Vector{UInt8}.

Base.seek — Method

seek(section::SectionRef, offset)

Seek to offset relative to section in the ObjectHandle that this SectionRef refers to

Base.seekstart — Method

seekstart(section::SectionRef)

Seek to the beginning of section in the ObjectHandle it was loaded from.

ObjectFile.canonical_rpaths — Method

canonical_rpaths(rpath::RPath)

Return a canonicalized list of paths that will be searched.

ObjectFile.deref — Method

deref(section::SectionRef)

Dereference the given SectionRef object to a Section.

ObjectFile.deref — Method

deref(seg::SegmentRef)

Dereference the given SegmentRef object to a Segment.

ObjectFile.endianness — Method

endianness(oh::ObjectHandle)

Returns the endianness of the given ObjectHandle (e.g. :LittleEndian)

ObjectFile.find_libraries — Method

find_libraries(oh::ObjectHandle)

Return a mapping from sonames to absolute paths, containing all the sonames declared as beeing needed by the given ObjectHandle. See the documentation for find_library(::RPath, ::String) and RPath for more details.

ObjectFile.find_library — Method

find_library(oh::ObjectHandle, soname::String)

Return the absolute path to the given soname, using the linker search path that the given ObjectHandle would use at runtime. See the documentation for find_library(::RPath, ::String) for more details.

ObjectFile.find_library — Method

find_library(rpath::RPath, soname::String)

Return the full path to a library, searching the given RPath, and then the default library search paths. This method takes the given soname and joins it to the end of every path within the given RPath, returning the resultant path if it exists, returning back the original soname if it doesn't.

ObjectFile.format_string — Method

format_string(::Type{H}) where {H <: ObjectHandle}

Return the string name of the given ObjectHandle, examples are "ELF", "MachO", "COFF", etc...

ObjectFile.getindex_ref — Method

getindex_ref(collection, offset, stride, T, ref_type, idx)

Given a collection, such as Sections, DynamicLinks, etc... use the given offset, stride, and T parameters to read in and construct a ref_type object located at index idx. Example invocation:

getindex_ref(
    sections,
    section_header_offset(oh),
    section_header_size(oh),
    section_header_type(oh),
    SectionRef,
    idx
)

ObjectFile.handle — Method

handle(rpath::RPath)

Return the handle that this RPath object refers to.

ObjectFile.handle — Method

handle(section::SectionRef)

Return the ObjectHandle this SectionRef belongs to. This method is SectionRef-only.

ObjectFile.handle — Method

handle(sections::Sections)

Return the ObjectHandle that this Sections object belongs to

ObjectFile.handle — Method

handle(seg::SegmentRef)

Return the ObjectHandle this SegmentRef belongs to. This method is SegmentRef-only.

ObjectFile.handle — Method

handle(s::StrTab)

Return the ObjectHandle this StrTab belongs to.

ObjectFile.handle — Method

handle(sym::SymbolRef)

Return the ObjectHandle that this SymbolRef belongs to.

ObjectFile.header — Method

header(oh::ObjectHandle)

Return the ObjectHandle's header object, whatever that may be for this particular object file format.

ObjectFile.iostream — Method

iostream(oh::ObjectHandle)

Returns the IOStream backing the ObjectHandle

ObjectFile.is64bit — Method

is64bit(oh::ObjectHandle)

Returns true if the given ObjectHandle represents a 64-bit object

ObjectFile.isdynamic — Method

isdynamic(oh::ObjectHandle)

Returns true if the given ObjectHandle makes use of dynamic linking.

ObjectFile.isexecutable — Method

isexecutable(oh::ObjectHandle)

Returns true if the given ObjectHandle represents an executable object

ObjectFile.isglobal — Method

isglobal(sym::SymtabEntry)

Return true if the given symbol is global

ObjectFile.islibrary — Method

islibrary(oh::ObjectHandle)

Returns true if the given ObjectHandle represents a shared library

ObjectFile.islocal — Method

islocal(sym::SymtabEntry)

Return true if the given symbol is local

ObjectFile.isrelocatable — Method

isrelocatable(oh::ObjectHandle)

Returns true if the given ObjectHandle represents a relocatable object file, e.g. an .o file as generated by gcc -c

ObjectFile.isundef — Method

isundef(sym::SymtabEntry)

Return true if the given symbol is undefined

ObjectFile.isweak — Method

isweak(sym::SymtabEntry)

Return true if the given symbol is weak

ObjectFile.mangle_section_name — Method

mangle_section_name(oh::ObjectHandle, name::AbstractString)

Turn a section name into the object-format specific naming convention, e.g. returning ".bss" for ELF/COFF files, and "__bss" for MachO files

ObjectFile.mangle_symbol_name — Method

mangle_symbol_name(oh::ObjectHandle, name::AbstractString)

Mangle a symbol name using the object-format specific naming convention, e.g. prefixing "_" for MachO files.

ObjectFile.path — Method

path(io::IO)

Try to guess the path of an IO object. If it cannot be guessed, returns the empty string.

ObjectFile.path — Method

path(oh::ObjectHandle)

Return the absolute path to the given ObjectHandle, if it was a file loaded from the local disk. If it was loaded from a general IOStream or in some other way such that the path is unknown or unknowable, return the empty string.

ObjectFile.read_struct — Method

read_struct(oh, type_func, size_func, name)

Given a Type, (such as ELFSection64), unpack() it from the given object and return it, throwing errors as appropriate, and skipping over any excess padding bytes as determined by type_func and size_func. Example invocation:

readstruct(oh, symtabentrytype, symtabentry_size, "Symbol Entry")

ObjectFile.readmeta — Method

readmeta(f::Function, file::AbstractString)

Do-block variant of readmeta(). Use via something like:

readmeta("libfoo.so") do f
    ...
end

ObjectFile.readmeta — Method

readmeta(io::IO)

Read an Object File out from an IOStream, guessing at the type of object within the stream by calling readmeta(io, T) for each T within ObjTypes, and returning the first that does not throw a MagicMismatch.

ObjectFile.readmeta — Method

readmeta(io::IO, ::ObjectHandle)

Read an Object File out from an IOStream. This is the first method you should call in order to manipulate object files.

ObjectFile.rpaths — Method

rpaths(rpath::RPath)

Return the list of paths that will be searched for shared libraries.

ObjectFile.section_address — Method

section_address(section::Section)

The address of the section in virtual memory.

ObjectFile.section_header_offset — Method

section_header_offset(oh::ObjectHandle)

Given an ObjectHandle, return the offset (in bytes) at which the sections start within the containing object file.

ObjectFile.section_header_size — Method

section_header_size(oh::ObjectHandle)

Given an ObjectHandle, return the size of a section header (used for reading in the sections header when trying to load a Section object or iterating over a Sections object)

ObjectFile.section_header_type — Method

section_header_type(oh::ObjectHandle)

Given an ObjectHandle, return the type of a section header (used for reading in the sections header when trying to load a Section object or iterating over a Sections object). For instance, for a 64-bit ELF file, this would return the type ELFSection64

ObjectFile.section_name — Method

section_name(section::Section)

Return the name of the given section as a string. In order to return a true name, it is necessary to perform a lookup within the object's string table, which cannot be done using just a Section object; use a SectionRef object instead if you need that. For sanity sake, this method will return a string, but the contents of the string may be something like the offset within the string table pointing to this Section's name, e.g. "@strtab.123"

ObjectFile.section_number — Method

section_number(section::SectionRef)

The index of the given section within the section header table.

ObjectFile.section_offset — Method

section_offset(section::Section)

The offset of the section in the file, in bytes

ObjectFile.section_size — Method

section_size(section::Section)

The size of the actual data contained in the section. This should exclude any padding mandated by the file format e.g. due to alignment rules

ObjectFile.segment_header_offset — Method

segment_header_offset(oh::ObjectHandle)

Given an ObjectHandle, return the offset (in bytes) at which the segments start within the containing object file.

ObjectFile.segment_header_size — Method

segment_header_size(oh::ObjectHandle)

Given an ObjectHandle, return the size of a segment header (used for reading in the segments header when trying to load a Segment object or iterating over a Segments object)

ObjectFile.segment_header_type — Method

segment_header_type(oh::ObjectHandle)

Given an ObjectHandle, return the type of a segment header (used for reading in the segments header when trying to load a Segment object or iterating over a Segments object). For instance, for a 64-bit ELF file, this would return the type ELFSegment64

ObjectFile.segment_name — Method

segment_name(seg::SegmentRef)

The name of the given Segment, returned as a string. This method often performs some kind of lookup within the string table of the object to get the full name of the segment.

ObjectFile.segment_number — Method

segment_number(seg::SegmentRef)

Return the index of the referred segment.

ObjectFile.strtab_lookup — Method

strtab_lookup(s::StrTab, index)

Reads a string from the given StrTab at index.

ObjectFile.symbol_name — Method

symbol_name(sym::SymbolRef)

Return the name of the given symbol as a string. This method often performs some kind of lookup within the string table of the object to get the full name of the symbol.

ObjectFile.symbol_name — Method

symbol_name(sym::SymtabEntry)

Return the name of the given section as a string. In order to return a true name, it is necessary to perform a lookup within the object's string table, which cannot be done using just a SymtabEntry object; use a SymbolRef object instead if you need that. For sanity sake, this method will return a string, but the contents of the string may be something like the offset within the string table pointing to this SymtabEntry's name, e.g. "@strtab.123"

ObjectFile.symbol_number — Method

symbol_number(sym::SymbolRef)

Return the number (index) of the given symbol.

ObjectFile.symbol_value — Method

symbol_value(sym::SymtabEntry)

Return the value of the given symbol

ObjectFile.symtab_entry_offset — Method

symtab_entry_offset(oh::ObjectHandle)

Given an ObjectHandle, return the offset (in bytes) at which the symbol table starts within the containing object file.

ObjectFile.symtab_entry_size — Method

symtab_entry_size(oh::ObjectHandle)

Given an ObjectHandle, return the size of a symbol table entry (used for reading in the symbol table when trying to load a SymtabEntry object or iterating over a Symbols object).

ObjectFile.symtab_entry_type — Method

symtab_entry_type(oh::ObjectHandle)

Given an ObjectHandle, return the type of a symbol table entry (used for reading in the symbol table when trying to load a SymtabEntry object or iterating over a Symbols object). For instance, for a 64-bit ELF file, this would return the type ELFSymtabEntry64

ObjectFile.DynamicLink — Type

DynamicLink

This type encapsulates the linkage of one object file to another. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

DynamicLink()

Accessors:

DynamicLinks()

handle()

path()

ObjectFile.DynamicLinks — Type

DynamicLinks

This type encapsulates the list of dynamic links within an object, holding a collection of DynamicLink objects. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

DynamicLinks()

Iteration

getindex()

lastindex()

iterate()

keys()

eltype()

Misc.

handle()

ObjectFile.MagicMismatch — Type

MagicMismatch

This is an error type used to denote that readmeta() was called on a file that does not contain the proper magic at the beginning for the type of object file that was attempting to be loaded.

ObjectFile.ObjectHandle — Type

ObjectHandle

The basic type that provides access to object files. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis. Note tha "must implement" is a bit of a misnomer, if an Object file does not have need of a certain piece of this API (e.g. COFF files have no concept of Segments), leaving that chunk of the API undefined will simply cause errors if a user attempts to use methods that use that part of the API (in the example above, an error will be thrown if the user calls Segments(oh) where oh <: COFFHandle).

Creation

readmeta()

IOStream-like operations:

seek()

seekstart()

skip()

startaddr()

iostream()

position()

read()

readuntil()

eof()

unpack()

Format-specific properties

endianness()

is64bit()

isrelocatable()

isexecutable()

islibrary()

isdynamic()

manglesectionname()

manglesymbolname()

handle()

header()

format_string()

Section properties

sectionheaderoffset()

sectionheadersize()

sectionheadertype()

Segment properties

segmentheaderoffset()

segmentheadersize()

segmentheadertype()

Symbol properties

symtabentryoffset()

symtabentrysize()

symtabentrytype()

Misc

path()

show()

find_library()

find_libraries()

ObjectFile.RPath — Type

RPath

This type encapsulates the search path used by an object file when looking for a shared library. This class enables not only looking at the path, but querying the path for matches for given library names. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

RPath()

Utility

handle()

RPath operations

rpaths()

canonical_rpaths()

find_library()

lastindex()

iterate()

ObjectFile.RPath — Method

RPath(oh::ObjectHandle)

Construct an RPath object from the given ObjectHandle.

ObjectFile.Section — Type

Section

An abstraction over the concept of a Section within an object file. Because many operations upon sections require global operations (access to the string table, knowledge of position within the file, etc...) some operations are defined only upon the SectionRef datatype. As a user, the SectionRef type should be the primary method of interacting with sections, as a developer adding new object file formats, some methods must support Sections, others must support only SectionRefs. Note that any method that works on a Section must also work with a SectionRef, see the @derefmethod macro for a convenient helper macro to generate SectionRef -> Section wrapper methods. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

read()

Utility:

deref()

IO-like operations:

read()

Format-specific properties:

section_name()

section_size()

section_offset()

section_address()

ObjectFile.SectionRef — Type

SectionRef

Provides a reference to a Section, along with a reference to the ObjectHandle this Section comes from. This should be the primary method by which users interact with sections inside object files. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis. Note that this overlaps heavily with the Section object API, this is by design as many of the methods are simply passthroughs to the underlying Section API calls for ease of use.

Creation:

SectionRef()

Utility

deref()

handle()

Sections()

IO-like operations:

read()

seekstart()

seek()

eof()

Format-specific properties:

section_name()

section_number()

section_type()

section_size()

section_offset()

section_address()

ObjectFile.SectionRef — Method

SectionRef(sections::Sections, section::Section, idx)

Construct a SectionRef object pointing to the given Section, which itself represents the idx'th section within the given Sections.

ObjectFile.Sections — Type

Sections

An abstraction over the concept of a collection of Section types within an object file. One can think of the Sections object containing the table of section headers within the object file, whereas the Section/SectionRef objects contain the actual section data itself. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

Sections()

Iteration

getindex()

lastindex()

length()

iterate()

keys()

eltype()

Search

findall()

findfirst()

Misc.

handle()

ObjectFile.Sections — Method

Sections(section::SectionRef)

Return the Sections collection this section belongs to.

ObjectFile.Segment — Type

Segment

An abstraction over the concept of a Segment within an object file. A Segment is a portion of an object file that is given instruction on its layout in virtual memory; this is in contrast to a Section, which delineates different portions of an object file on disk. ELF files have the strictest separation here, with a single executable file containing multiple Segment and Section objects, with Sections being assigned to one or more Segments for virtual memory placement. Mach-O files typically have two Segments, one called __TEXT, one called __DATA. COFF files do not have Segment.

Just like with Section objects, many operations upon segments require global operations (access to the string table, knowledge of position within the file, etc...) which causes some operations to be defined only upon the SegmentRef datatype. As a user, the SegmentRef type should be the primary method of interacting with segments, as a developer adding new object file formats, some methods must support Segments, others must support only SegmentRefs. Note that any method that works on a Segment must also work with a SegmentRef, see the @derefmethod macro for a convenient helper macro to generate SegmentRef -> Section wrapper methods. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

read()

Utility:

deref()

Format-specific properties:

segment_name()

segment_offset()

segmentfilesize()

segmentmemorysize()

segment_address()

ObjectFile.SegmentRef — Type

SegmentRef

Provides a reference to a Segment, along with a reference to the ObjectHandle this Segment comes from. This should be the primary method by which users interact with segments inside object files. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis. Note that this overlaps heavily with the Segment object API, this is by design as many of the methods are simply passthroughs to the underlying Segment API calls for ease of use.

Creation:

SegmentRef()

Utility

deref()

Segments()

handle()

Format-specific properties:

segment_name()

segment_number()

segment_offset()

segmentfilesize()

segmentmemorysize()

segment_address()

ObjectFile.Segments — Type

Segments

An abstraction over the concept of a collection of Segment types within an object file. One can think of the Segments object containing the table of segment headers within the object file, whereas the Segment/SegmentRef objects contain the actual segment data itself. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

Segments()

Iteration

getindex()

lastindex()

length()

iterate()

keys()

eltype()

Search

findall()

findfirst()

Misc.

handle()

ObjectFile.Segments — Method

Segments(seg::SegmentRef)

Return the Segments collection this Segment belongs to.

ObjectFile.StrTab — Type

StrTab

This type encapsulates a string table within an object file, enabling queries against the string table for symbol names, section names, etc... The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

StrTab()

Accessors

handle()

strtab_lookup()

ObjectFile.SymbolRef — Type

SymbolRef

Provides a reference to a SymtabEntry, along with a reference to the ObjectHandle this SymtabEntry comes from. This should be the primary method by which users interact with symbols inside object files. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis. Note that this overlaps heavily with the SymtabEntry object API, this is by design as many of the methods are simply passthroughs to the underlying SymtabEntry API calls for ease of use.

Creation:

SymbolRef()

Util:

deref()

Symbols()

handle()

Properties:

symbol_number()

symbol_name()

symbol_value()

isundef()

ObjectFile.SymbolRef — Method

SymbolRef(symbols::Symbols, sym::SymtabEntry, idx::UInt32)

Construct a SymbolRef object pointing to the given SymtabEntry, which itself represents the idx'th symbol within the given Symbols collection.

ObjectFile.Symbols — Type

Symbols

An abstraction over the concept of a collection of symbol (SymtabEntry) types within an object file. One can think of the Symbols object containing the table of symbols within the object file, whereas the SymtabEntry/SymbolRef objects contain the actual symbol data itself. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation

Symbols()

Iteration

getindex()

lastindex()

length()

iterate()

eltype()

Misc.

handle()

ObjectFile.Symbols — Method

Symbols(sym::SymbolRef)

Return the Symbols object that this SymbolRef belongs to.

ObjectFile.SymtabEntry — Type

SymtabEntry

An abstraction over the concept of a symbol within an object file. This type does not use the Symbol name as this would conflict with the builtin Julia Symbol type, so the name SymtabEntry is used instead. As a user, the SymbolRef type should be the primary method of interacting with symbols, as a developer adding new object file formats, some methods must support SymtabEntrys, others must support only SymbolRefs. Note that any method that works on a SymtabEntry must also work with a SymbolRef, see the @derefmethod macro for a convenient helper macro to generate SymbolRef -> SymtabEntry wrapper methods. The list of available API operations is given below, with methods that subclasses must implement marked in emphasis:

Creation:

SymtabEntry()

Util:

deref()

Properties:

symbol_name()

symbol_value()

isundef()

isglobal()

islocal()

isweak()

ObjectFile.fixed_string — Type

fixed_string{T}

A fixed-byte string, stored as an integer type (e.g. T = UInt128, or T = UInt64) but displayed and treated as a string.

Base.unsafe_string — Method

unsafe_string(io, max_len::Integer)

Read in a null-terminated string, stopping with a maximum length of max_len.

Base.unsafe_string — Method

unsafe_string(x::fixed_string)

Convert a fixed_string object to a native-Julia String

ObjectFile.show_collection — Method

show_collection(io, collection, ::Type{H<:ObjectHandle})

Given a collection-like object, (Symbols, DynamicLinks, ``)

ObjectFile.startaddr — Method

startaddr(oh::ObjectHandle)

Returns the offset within the underlying IOStream at which this ObjectHandle is located.